インストールのテーブルDROP時に画面が乱れる症状を回避

+++ data/class/helper/SC_Helper_Session.php
@@ -12,7 +12,9 @@
   * @author LOCKON CO.,LTD.
   * @version $Id$
   */
- class SC_Helper_Session {
+class SC_Helper_Session {
+
+    var $objDb;
 
      // }}}
      // {{{ constructor
@@ -23,15 +25,13 @@
       * 各関数をセッションハンドラに保存する
       */
      function SC_Helper_Session() {
-         $objDb = new SC_Helper_DB_Ex();
-         if($objDb->sfTabaleExists("dtb_session")) {
-             session_set_save_handler(array(&$this, "sfSessOpen"),
-                                      array(&$this, "sfSessClose"),
-                                      array(&$this, "sfSessRead"),
-                                      array(&$this, "sfSessWrite"),
-                                      array(&$this, "sfSessDestroy"),
-                                      array(&$this, "sfSessGc"));
-         }
+         $this->objDb = new SC_Helper_DB_Ex();
+         session_set_save_handler(array(&$this, "sfSessOpen"),
+                                  array(&$this, "sfSessClose"),
+                                  array(&$this, "sfSessRead"),
+                                  array(&$this, "sfSessWrite"),
+                                  array(&$this, "sfSessDestroy"),
+                                  array(&$this, "sfSessGc"));
      }
 
      // }}}
@@ -64,12 +64,13 @@
       * @return string セッションデータの値
       */
      function sfSessRead($id) {
+         if (!$this->objDb->sfTabaleExists("dtb_session")) return '';
          $objQuery = new SC_Query();
          $arrRet = $objQuery->select("sess_data", "dtb_session", "sess_id = ?", array($id));
          if (empty($arrRet)) {
-             return null;
+             return '';
          } else {
-             return($arrRet[0]['sess_data']);
+             return $arrRet[0]['sess_data'];
          }
      }
 
@@ -82,6 +83,7 @@
       */
      function sfSessWrite($id, $sess_data)
      {
+         if (!$this->objDb->sfTabaleExists("dtb_session")) return false;
          $objQuery = new SC_Query();
          $count = $objQuery->count("dtb_session", "sess_id = ?", array($id));
          $sqlval = array();
@@ -112,6 +114,7 @@
       * @return bool セッションを正常に破棄した場合 true
       */
      function sfSessDestroy($id) {
+         if (!$this->objDb->sfTabaleExists("dtb_session")) return false;
          $objQuery = new SC_Query();
          $objQuery->delete("dtb_session", "sess_id = ?", array($id));
          return true;
@@ -126,6 +129,7 @@
       */
      function sfSessGc($maxlifetime) {
          // MAX_LIFETIME以上更新されていないセッションを削除する。
+         if (!$this->objDb->sfTabaleExists("dtb_session")) return false;
          $objQuery = new SC_Query();
          $where = "update_date < current_timestamp + '-". MAX_LIFETIME . " secs'";
          $objQuery->delete("dtb_session", $where);
